wsd: jail: introduce the Template directory
SysTemplate and LoTemplate are used to create the
final jail directory. This is done either by bind-mounting,
or, when mounting is unavailable, by hard-linking.

Unfortunately when bind-mounting is not possible,
we must resort to hard-linking, which can fail in
multiple scenarios, forcing a very slow and disk-space
heavy copying.

Hard-linking fails when the SysTemplate or LoTemplate
directory is on a different filesystem/mount-point,
or when filesystem namespaces are used. Another
reason is when either of these are owned by root
(or lool doesn't have write access to them).

To avoid all these problems, we create a Template
directory in Child-Root that is almost completely
ready for jailing (it only needs the random nodes
to be complete, which is done by kit, since loolwsd
doesn't have cap_mknod).

Since Template is created in Child-Root it can be
hard-linked to its files, and creating a jails
very fast is guaranteed. Of course ideally bind-mounting
will be used, which is faster and more secure (as
it's read-only).

Change-Id: I4058113ca42e6889129378218e3509384f3962d9
Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>