wsd: reset the modified flag when saving unmodified doc
In case saving can't really be done (f.e. the doc
is unmodified and there is nothing to save), we
must recognize that we are in fact up-to-date
and there is nothing to do.

Unfortunately this wasn't the case. When we issued
a save request, we only considered it done when it
succeeded. When it failed, we assumed we will
have to try again, until we succeed.

This logic is valid in most normal cases, except
when the document is unmodified (but we think
it *might* be modified).

Here we make sure that we update the last
save attempt so we don't think there might
be some modifications, which we track by
looking at the last activity timestamp compared
to the last save attempt.

This makes it possible to sync-up with Core
when it tells us that there was nothing to save
as the document isn't modified (since the last
save), and so we stop retrying to save an
unmodified document endlessly, which can hold
off an unloading document until timeout, and
even then log a scary warning that modifications
may be lost etc.).

Change-Id: Ia9a91dbc46b2e5b8e08e7caba6472b2315de7afa
Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>